Worst-to-best: split storyLabEngine's five mixed concerns, dedupe collapseWhitespace, add guidance unit tests
storyLabEngine.ts was the largest file in the repo (1,726 lines) and had
never had a structural pass, mixing route orchestration, a ~700-line
prompt-guidance subsystem, state-merge logic, response-shaping, and string
utilities in one file. Extracts the prompt-guidance system into
continuationGuidance.ts and state-merge logic into storyStateBuilder.ts,
consolidates three independent collapseWhitespace implementations into
api/_lib/utils/whitespace.ts, and adds direct unit tests for the
previously untested activation-scoring/pressure-selection logic.

storyLabEngine.ts: 1,726 -> 676 lines. Pure extraction + dedup, no behavior
change.
