+++ package.json
@@ -49,6 +49,7 @@
     "test:story-lab-evaluate-route": "tsx tests/story-lab-evaluate-route.test.ts",
     "test:story-lab-continuity-prompt": "tsx tests/story-lab-continuity-prompt.test.ts",
     "test:story-lab-continuity-merge": "tsx tests/story-lab-continuity-merge.test.ts",
+    "test:story-lab-continuation-guidance": "tsx tests/story-lab-continuation-guidance.test.ts",
     "test:story-lab-job-routes": "tsx tests/story-lab-job-routes.test.ts",
     "test:story-lab-job-store-port": "tsx tests/story-lab-job-store-port.test.ts",
     "test:story-lab-job-store-config": "tsx tests/story-lab-job-store-config.test.ts",
@@ -65,7 +66,7 @@
     "test:recovery-finish-check": "node --test tests/recovery-finish-check.test.mjs",
     "test:recovery-open-pr-summary": "node --test tests/recovery-open-pr-summary.test.mjs",
     "test:story-quality": "tsx tests/story-quality-evals.test.ts",
-    "test:all": "npm run test:story && npm run test:xai-fast-path-review && npm run test:verify-ai-fixes && npm run test:tropes && npm run test:cliffhangers && npm run test:story-continuation-cliffhanger-variety && npm run test:api-key-auth && npm run test:api-access-control && npm run test:export-service && npm run test:image-service && npm run test:logger-recent-logs && npm run test:story-route-contracts && npm run test:story-lab-state && npm run test:story-lab-state-store && npm run test:story-lab-blueprint-parser && npm run test:story-text-blocks && npm run test:story-lab-theme-seed-usage && npm run test:event-stream-retry && npm run test:streaming-progress-estimate && npm run test:text-excerpt && npm run test:html-document-download && npm run test:story-lab-real-engine && npm run test:story-lab-route-status && npm run test:story-lab-stream-genesis && npm run test:story-service-prompt-guards && npm run test:story-content-analysis && npm run test:story-lab-auth && npm run test:story-lab-configured-auth && npm run test:story-lab-clerk-auth && npm run test:story-lab-storage-port && npm run test:story-lab-cloud-schema && npm run test:story-lab-cloud-storage-config && npm run test:story-lab-neon-executor && npm run test:story-lab-cloud-schema-migration && npm run test:story-lab-cloud-db-readiness && npm run test:story-lab-profile-contracts && npm run test:story-lab-profile-store && npm run test:story-lab-privacy-contracts && npm run test:story-lab-account-routes && npm run test:story-lab-evaluate-route && npm run test:story-lab-continuity-prompt && npm run test:story-lab-continuity-merge && npm run test:story-lab-job-routes && npm run test:story-lab-job-store-port && npm run test:story-lab-job-store-config && npm run test:story-download-filename && npm run test:proving-grounds-production-prompt && npm run test:story-generator-route-splitting && npm run test:story-generator-component-style-budget && npm run test:review-thread-audit-options && npm run test:api-response-status && npm run test:method-not-allowed-allow-header && npm run test:request-correlation-id && npm run test:express-api-routes && npm run test:recovery-finish-check && npm run test:recovery-open-pr-summary && npm run test:story-quality",
+    "test:all": "npm run test:story && npm run test:xai-fast-path-review && npm run test:verify-ai-fixes && npm run test:tropes && npm run test:cliffhangers && npm run test:story-continuation-cliffhanger-variety && npm run test:api-key-auth && npm run test:api-access-control && npm run test:export-service && npm run test:image-service && npm run test:logger-recent-logs && npm run test:story-route-contracts && npm run test:story-lab-state && npm run test:story-lab-state-store && npm run test:story-lab-blueprint-parser && npm run test:story-text-blocks && npm run test:story-lab-theme-seed-usage && npm run test:event-stream-retry && npm run test:streaming-progress-estimate && npm run test:text-excerpt && npm run test:html-document-download && npm run test:story-lab-real-engine && npm run test:story-lab-route-status && npm run test:story-lab-stream-genesis && npm run test:story-service-prompt-guards && npm run test:story-content-analysis && npm run test:story-lab-auth && npm run test:story-lab-configured-auth && npm run test:story-lab-clerk-auth && npm run test:story-lab-storage-port && npm run test:story-lab-cloud-schema && npm run test:story-lab-cloud-storage-config && npm run test:story-lab-neon-executor && npm run test:story-lab-cloud-schema-migration && npm run test:story-lab-cloud-db-readiness && npm run test:story-lab-profile-contracts && npm run test:story-lab-profile-store && npm run test:story-lab-privacy-contracts && npm run test:story-lab-account-routes && npm run test:story-lab-evaluate-route && npm run test:story-lab-continuity-prompt && npm run test:story-lab-continuity-merge && npm run test:story-lab-continuation-guidance && npm run test:story-lab-job-routes && npm run test:story-lab-job-store-port && npm run test:story-lab-job-store-config && npm run test:story-download-filename && npm run test:proving-grounds-production-prompt && npm run test:story-generator-route-splitting && npm run test:story-generator-component-style-budget && npm run test:review-thread-audit-options && npm run test:api-response-status && npm run test:method-not-allowed-allow-header && npm run test:request-correlation-id && npm run test:express-api-routes && npm run test:recovery-finish-check && npm run test:recovery-open-pr-summary && npm run test:story-quality",
     "smoke:grok": "tsx tests/grok-multi-agent-smoke.test.ts",
     "smoke:grok-multi-agent": "tsx tests/grok-multi-agent-smoke.test.ts",
     "smoke:story-lab-ui": "node scripts/recovery/story-lab-browser-smoke.mjs",
